add soft delete add comments for sql make sure deleted uses "1" add deleted to kvobject add permanent and soft delete add sql for "deleted" - soft delete

+++ src/com/migrate/service/DataService.java
@@ -17,49 +17,59 @@ import com.migrate.storage.ObjectStore;
 
 @Component("dataService")
 public class DataService {
-	private static final String NAMESPACE = "__data";
-	private static org.apache.log4j.Logger log = Logger.getLogger(SchemaService.class);
+    private static final String NAMESPACE = "__data";
+    private static org.apache.log4j.Logger log = Logger.getLogger(SchemaService.class);
 
     @Autowired
-	@Qualifier(value = "objectStore")
-	private ObjectStore store;
+    @Qualifier(value = "objectStore")
+    private ObjectStore store;
 
-	@Autowired
-	@Qualifier(value = "luceneIndexService")
-	private LuceneIndexService luceneIndexService;
+    @Autowired
+    @Qualifier(value = "luceneIndexService")
+    private LuceneIndexService luceneIndexService;
 
-	public GenericMap getObject(String className, String id) throws IOException {
-		return store.get(NAMESPACE, id, className, GenericMap.class);	
-	}
-	
-	public void storeObject(GenericMap data) throws IOException {
-		data.setWd_namespace(NAMESPACE);
-		luceneIndexService.updateIndex(data);
-		store.update(data);
-	}
-	
-	public void deleteObject(String className, String id) throws IOException {
-        deleteObject(className, id, false);
+    public GenericMap getObject(String className, String id) throws IOException {
+        return store.get(NAMESPACE, id, className, GenericMap.class);
     }
 
-    public void deleteObject(String className, String id, boolean isHard) throws IOException {
-        // TODO: ok drop the index, but still
-		luceneIndexService.deleteIndex(className, id);
-		store.delete(NAMESPACE,  id);
-	}
-	
-	public void createObject(GenericMap data) throws IOException {
-		log.info("calling createObject");
-		luceneIndexService.updateIndex(data);
-		data.setWd_namespace(NAMESPACE);
-		store.create(data);
-	}
-	
-	public List<GenericMap> luceneSearch(String className, String queryStr) throws ParseException, IOException {
-		return luceneIndexService.search(className, queryStr);
-	}
-	
-	public List<GenericMap> find(String className, long time, int start, int numMatches) throws ParseException, IOException {
-		return store.findChanged(NAMESPACE, className, time, start, numMatches);
-	}
+    public void storeObject(GenericMap data) throws IOException {
+        data.setWd_namespace(NAMESPACE);
+        luceneIndexService.updateIndex(data);
+        store.update(data);
+    }
+
+    public void deleteObject(String className, String id) throws IOException {
+        deleteObject(className, id, false, 0L);
+    }
+
+    public void softDeleteObject(String className, String id, long now) throws IOException {
+        deleteObject(className, id, false, now);
+    }
+
+    public void permanentDelete(String className, String id) throws IOException {
+        deleteObject(className, id, true, 0L);
+    }
+
+    public void deleteObject(String className, String id, boolean isPermanent, long now) throws IOException {
+        // TODO: ok drop the index, but still keep the deleted object?
+        if (isPermanent) {
+            luceneIndexService.deleteIndex(className, id);
+        }
+        store.delete(NAMESPACE, id, isPermanent, now);
+    }
+
+    public void createObject(GenericMap data) throws IOException {
+        log.info("calling createObject");
+        luceneIndexService.updateIndex(data);
+        data.setWd_namespace(NAMESPACE);
+        store.create(data);
+    }
+
+    public List<GenericMap> luceneSearch(String className, String queryStr) throws ParseException, IOException {
+        return luceneIndexService.search(className, queryStr);
+    }
+
+    public List<GenericMap> find(String className, long time, int start, int numMatches) throws ParseException, IOException {
+        return store.findChanged(NAMESPACE, className, time, start, numMatches);
+    }
 }
